# [Perf] end-to-end benchmark, configurable prompt-len and gen-len