We need to have a rigorous e2e speed benchmark to compare different implementations.

Also, we want to see how dynamic len affects perf.
 
Example usage:

```
~/mlc-llm (e2e-bench ✔) ./build/mlc_chat_cli --local-id Llama-2-7b-chat-hf-q4f16_1 --evaluate --eval-prompt-len 1024 --eval-gen-len=256
Use MLC config: "/home/bohan/mlc-llm/dist/Llama-2-7b-chat-hf-q4f16_1/params/mlc-chat-config.json"
Use model weights: "/home/bohan/mlc-llm/dist/Llama-2-7b-chat-hf-q4f16_1/params/ndarray-cache.json"
Use model library: "/home/bohan/mlc-llm/dist/Llama-2-7b-chat-hf-q4f16_1/Llama-2-7b-chat-hf-q4f16_1-cuda.so"
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:678: encoding-time=466.727ms, 
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 0] decoding-time=8.62598ms tok/s: 115.929.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 1] decoding-time=8.57168ms tok/s: 116.295.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 2] decoding-time=8.54583ms tok/s: 116.534.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 3] decoding-time=8.53918ms tok/s: 116.677.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 4] decoding-time=8.36347ms tok/s: 117.244.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 5] decoding-time=8.3658ms tok/s: 117.62.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 6] decoding-time=8.34194ms tok/s: 117.937.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 7] decoding-time=8.33219ms tok/s: 118.193.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 8] decoding-time=8.34095ms tok/s: 118.379.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 9] decoding-time=8.34395ms tok/s: 118.524.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 10] decoding-time=8.34382ms tok/s: 118.643.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 11] decoding-time=8.32971ms tok/s: 118.76.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 12] decoding-time=8.34704ms tok/s: 118.839.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 13] decoding-time=8.35365ms tok/s: 118.901.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 14] decoding-time=8.36279ms tok/s: 118.946.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 15] decoding-time=8.3497ms tok/s: 118.997.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 16] decoding-time=8.36659ms tok/s: 119.027.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 17] decoding-time=8.36766ms tok/s: 119.054.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 18] decoding-time=8.3618ms tok/s: 119.082.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 19] decoding-time=8.3597ms tok/s: 119.109.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 20] decoding-time=8.36502ms tok/s: 119.13.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 21] decoding-time=8.37303ms tok/s: 119.143.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 22] decoding-time=8.37585ms tok/s: 119.154.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 23] decoding-time=8.36937ms tok/s: 119.168.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 24] decoding-time=8.38352ms tok/s: 119.172.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 25] decoding-time=8.37848ms tok/s: 119.179.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 26] decoding-time=8.3861ms tok/s: 119.182.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 27] decoding-time=8.37894ms tok/s: 119.188.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 28] decoding-time=8.3883ms tok/s: 119.189.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 29] decoding-time=8.3965ms tok/s: 119.185.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 30] decoding-time=8.39461ms tok/s: 119.183.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 31] decoding-time=8.39209ms tok/s: 119.183.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 32] decoding-time=8.40086ms tok/s: 119.178.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 33] decoding-time=8.39503ms tok/s: 119.177.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 34] decoding-time=8.40837ms tok/s: 119.169.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 35] decoding-time=8.38703ms tok/s: 119.171.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 36] decoding-time=8.40495ms tok/s: 119.166.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 37] decoding-time=8.4067ms tok/s: 119.16.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 38] decoding-time=8.4095ms tok/s: 119.154.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 39] decoding-time=8.40512ms tok/s: 119.149.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 40] decoding-time=8.41326ms tok/s: 119.142.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 41] decoding-time=8.41499ms tok/s: 119.135.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 42] decoding-time=8.41845ms tok/s: 119.127.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 43] decoding-time=8.40723ms tok/s: 119.123.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 44] decoding-time=8.4243ms tok/s: 119.113.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 45] decoding-time=8.42402ms tok/s: 119.105.
[18:04:46] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 46] decoding-time=8.42426ms tok/s: 119.096.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 47] decoding-time=8.4272ms tok/s: 119.087.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 48] decoding-time=8.42424ms tok/s: 119.079.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 49] decoding-time=8.42139ms tok/s: 119.073.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 50] decoding-time=8.42624ms tok/s: 119.065.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 51] decoding-time=8.44127ms tok/s: 119.053.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 52] decoding-time=8.44137ms tok/s: 119.042.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 53] decoding-time=8.44311ms tok/s: 119.031.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 54] decoding-time=8.44955ms tok/s: 119.018.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 55] decoding-time=8.44172ms tok/s: 119.008.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 56] decoding-time=8.45729ms tok/s: 118.995.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 57] decoding-time=8.45781ms tok/s: 118.982.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 58] decoding-time=8.45759ms tok/s: 118.969.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 59] decoding-time=8.44973ms tok/s: 118.958.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 60] decoding-time=8.45829ms tok/s: 118.946.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 61] decoding-time=8.46753ms tok/s: 118.933.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 62] decoding-time=8.46236ms tok/s: 118.92.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 63] decoding-time=8.46057ms tok/s: 118.909.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 64] decoding-time=8.46908ms tok/s: 118.896.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 65] decoding-time=8.48096ms tok/s: 118.881.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 66] decoding-time=8.47396ms tok/s: 118.868.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 67] decoding-time=8.47002ms tok/s: 118.856.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 68] decoding-time=8.474ms tok/s: 118.844.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 69] decoding-time=8.47913ms tok/s: 118.831.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 70] decoding-time=8.48814ms tok/s: 118.816.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 71] decoding-time=8.48507ms tok/s: 118.803.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 72] decoding-time=8.48859ms tok/s: 118.789.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 73] decoding-time=8.48817ms tok/s: 118.776.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 74] decoding-time=8.489ms tok/s: 118.762.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 75] decoding-time=8.49497ms tok/s: 118.749.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 76] decoding-time=8.5036ms tok/s: 118.733.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 77] decoding-time=8.50247ms tok/s: 118.719.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 78] decoding-time=8.50801ms tok/s: 118.704.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 79] decoding-time=8.49677ms tok/s: 118.691.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 80] decoding-time=8.50331ms tok/s: 118.678.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 81] decoding-time=8.51596ms tok/s: 118.662.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 82] decoding-time=8.50749ms tok/s: 118.649.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 83] decoding-time=8.50216ms tok/s: 118.636.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 84] decoding-time=8.51635ms tok/s: 118.622.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 85] decoding-time=8.51969ms tok/s: 118.607.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 86] decoding-time=8.52236ms tok/s: 118.592.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 87] decoding-time=8.52171ms tok/s: 118.578.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 88] decoding-time=8.53198ms tok/s: 118.562.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 89] decoding-time=8.53193ms tok/s: 118.547.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 90] decoding-time=8.53302ms tok/s: 118.532.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 91] decoding-time=8.52908ms tok/s: 118.518.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 92] decoding-time=8.54296ms tok/s: 118.502.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 93] decoding-time=8.54814ms tok/s: 118.486.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 94] decoding-time=8.54201ms tok/s: 118.471.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 95] decoding-time=8.53598ms tok/s: 118.457.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 96] decoding-time=8.55239ms tok/s: 118.441.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 97] decoding-time=8.55198ms tok/s: 118.425.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 98] decoding-time=8.55802ms tok/s: 118.409.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 99] decoding-time=8.55458ms tok/s: 118.394.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 100] decoding-time=8.56192ms tok/s: 118.378.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 101] decoding-time=8.56686ms tok/s: 118.361.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 102] decoding-time=8.57701ms tok/s: 118.344.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 103] decoding-time=8.56745ms tok/s: 118.328.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 104] decoding-time=8.57315ms tok/s: 118.312.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 105] decoding-time=8.58279ms tok/s: 118.294.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 106] decoding-time=8.58362ms tok/s: 118.277.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 107] decoding-time=8.57195ms tok/s: 118.262.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 108] decoding-time=8.59337ms tok/s: 118.245.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 109] decoding-time=8.58407ms tok/s: 118.228.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 110] decoding-time=8.59306ms tok/s: 118.212.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 111] decoding-time=8.59025ms tok/s: 118.195.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 112] decoding-time=8.58831ms tok/s: 118.179.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 113] decoding-time=8.58642ms tok/s: 118.164.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 114] decoding-time=8.59171ms tok/s: 118.148.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 115] decoding-time=8.58546ms tok/s: 118.134.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 116] decoding-time=8.59789ms tok/s: 118.118.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 117] decoding-time=8.60197ms tok/s: 118.102.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 118] decoding-time=8.60342ms tok/s: 118.086.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 119] decoding-time=8.59134ms tok/s: 118.072.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 120] decoding-time=8.6046ms tok/s: 118.056.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 121] decoding-time=8.6143ms tok/s: 118.04.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 122] decoding-time=8.61152ms tok/s: 118.024.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 123] decoding-time=8.61261ms tok/s: 118.008.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 124] decoding-time=8.62341ms tok/s: 117.992.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 125] decoding-time=8.62182ms tok/s: 117.975.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 126] decoding-time=8.63075ms tok/s: 117.958.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 127] decoding-time=8.6227ms tok/s: 117.943.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 128] decoding-time=8.64334ms tok/s: 117.925.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 129] decoding-time=8.65277ms tok/s: 117.906.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 130] decoding-time=8.6485ms tok/s: 117.889.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 131] decoding-time=8.6347ms tok/s: 117.873.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 132] decoding-time=8.64937ms tok/s: 117.855.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 133] decoding-time=8.64456ms tok/s: 117.839.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 134] decoding-time=8.64683ms tok/s: 117.822.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 135] decoding-time=8.6509ms tok/s: 117.806.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 136] decoding-time=8.65486ms tok/s: 117.789.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 137] decoding-time=8.65543ms tok/s: 117.772.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 138] decoding-time=8.66348ms tok/s: 117.755.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 139] decoding-time=8.64457ms tok/s: 117.74.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 140] decoding-time=8.66427ms tok/s: 117.723.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 141] decoding-time=8.68044ms tok/s: 117.705.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 142] decoding-time=8.67032ms tok/s: 117.688.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 143] decoding-time=8.66121ms tok/s: 117.672.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 144] decoding-time=8.6741ms tok/s: 117.655.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 145] decoding-time=8.67395ms tok/s: 117.639.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 146] decoding-time=8.67793ms tok/s: 117.622.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 147] decoding-time=8.68693ms tok/s: 117.605.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 148] decoding-time=8.68997ms tok/s: 117.587.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 149] decoding-time=8.69158ms tok/s: 117.57.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 150] decoding-time=8.68371ms tok/s: 117.554.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 151] decoding-time=8.68213ms tok/s: 117.538.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 152] decoding-time=8.69337ms tok/s: 117.521.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 153] decoding-time=8.70453ms tok/s: 117.504.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 154] decoding-time=8.69279ms tok/s: 117.487.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 155] decoding-time=8.68424ms tok/s: 117.472.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 156] decoding-time=8.71422ms tok/s: 117.454.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 157] decoding-time=8.71336ms tok/s: 117.437.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 158] decoding-time=8.71663ms tok/s: 117.42.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 159] decoding-time=8.70626ms tok/s: 117.403.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 160] decoding-time=8.71512ms tok/s: 117.386.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 161] decoding-time=8.71474ms tok/s: 117.37.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 162] decoding-time=8.7248ms tok/s: 117.352.
[18:04:47] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 163] decoding-time=8.71052ms tok/s: 117.336.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 164] decoding-time=8.72413ms tok/s: 117.32.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 165] decoding-time=8.73335ms tok/s: 117.302.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 166] decoding-time=8.72583ms tok/s: 117.286.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 167] decoding-time=8.72855ms tok/s: 117.269.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 168] decoding-time=8.74532ms tok/s: 117.251.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 169] decoding-time=8.73997ms tok/s: 117.234.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 170] decoding-time=8.74851ms tok/s: 117.217.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 171] decoding-time=8.73441ms tok/s: 117.201.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 172] decoding-time=8.75298ms tok/s: 117.183.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 173] decoding-time=8.76139ms tok/s: 117.165.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 174] decoding-time=8.76136ms tok/s: 117.147.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 175] decoding-time=8.74773ms tok/s: 117.131.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 176] decoding-time=8.74726ms tok/s: 117.115.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 177] decoding-time=8.75704ms tok/s: 117.098.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 178] decoding-time=8.76133ms tok/s: 117.081.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 179] decoding-time=8.75137ms tok/s: 117.065.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 180] decoding-time=8.7673ms tok/s: 117.048.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 181] decoding-time=8.76698ms tok/s: 117.031.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 182] decoding-time=8.76825ms tok/s: 117.014.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 183] decoding-time=8.76375ms tok/s: 116.998.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 184] decoding-time=8.7661ms tok/s: 116.982.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 185] decoding-time=8.77673ms tok/s: 116.965.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 186] decoding-time=9.09489ms tok/s: 116.925.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 187] decoding-time=8.77153ms tok/s: 116.909.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 188] decoding-time=8.78953ms tok/s: 116.892.
[18:04:48] /home/bohan/mlc-llm/cpp/llm_chat.cc:689: [i: 189] decoding-time=8.78098ms tok/s: 116.876.
…